requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
attrs==25.3.0
bcrypt==4.3.0
black==24.4.2
certifi==2025.4.26
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.110.0
fastapi-cache2==0.2.1
flake8==7.0.0
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
motor==3.1.2
multidict==6.4.4
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.11.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.2.0
pymongo==4.3.3
pytest==8.2.0
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
